examples/research_projects/retroprompt/RE_task/requirements.txt
numpy>=1.20.3
tokenizers==0.10.3
pytorch_lightning>=1.3.1
regex==2021.4.4
torch==1.7.1+cu110
transformers==4.7.0
tqdm==4.49.0
activations==0.1.0
dataclasses==0.8
file_utils==0.0.1
flax==0.3.4
PyYAML==5.4.1
utils==1.0.1
